Initial commit
This commit includes the script that trains an LSTM model to predict
next-day's price of EUR/USD Forex pair and tests it on an unseen part of
the data. Also included is a .csv file with data for this pair from 2009
to 2019. Finally, there are plots of several models' behavior under
different hyperparameters.